requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
django-extensions==3.2.3
whitenoise==6.6.0
celery==5.3.4
redis==5.0.1